+++ Fig/Include/FigApi.hpp
@@ -41,7 +41,7 @@ FigCode __cdecl execute(__in OperationType operation,
 
 FigCode __cdecl status(__in OperationId id, __out ExecutionStatus* status, __out FigSpecificCode* specific_code);
 
-FigCode __cdecl take(__in OperationId id, __out uint8_t** buffer, __out uint32_t* buffer_size);
+FigCode __cdecl take(__in OperationId id, __out uint8_t** buffer, __out size_t* buffer_size);
 
 FigCode __cdecl free_buffer(__in uint8_t* buffer, __in uint32_t buffer_size);
